DataTypes 0.11, Analyser 0.27, HandTypeCheckers 0.4.2
Moved the addFrame and newPlayer functions from Analyser to DataTypes.
Introduced the new Deck ADT and implemented it in all relevant functions
in Analyser through the new function newDeck.
NOTE: From here onwards the Deck in the Frames includes all the player's
hands and the discarded cards; because of this the discard function and
some instances of cards in deck decrease have been removed.
HandTypeCheckers: as expected, reverted back to previous versions of all
checkers; they are much more legible, which is preferable to the
stand-alone-ness of the last ones, which was not really super-useful.