GeneralFunctions 0.2
Resolved numerical casting issue: made the previous version of choose
chooseUnsafe, and made the new version safe.
Now countRoyalFlush' works perfectly, however consider making the second
element of the return type an Int instead of [Int] if it turns out that
when not considering the cards that other players already have (which
cannot be known (usually)) the probabilities of each instance of a
possible HandType are all the same (resulting in a list of equal
values).