HandRankings 0.16
Remade the first two count functions in order to cater for empty list
cases.
Drafted a more abstract version; will have to decide which parts should
not be abstracted (like, perhaps, the CardSet value in the
HandTypeCount).
Renamed the fromGSV and fromGVS functions to fromSVG and fromVSG
(fromValuesSuitsGrouped and fromSuitsValuesGrouped).