HandCounters 0.14, QuickCheckStuff 0.6
handCounters: made all count and related functions (including the
testing ones) take in a list of Ints representing the target number of
cards and returning each corresponding HandTypeCount; made a simple
parallel version of checkAllHtCs.
QuickCheckStuff: changed the options from quick and verbose to normal
and parallel, the second one using the parallel function above.; it is
consistently faster.
DataTypes: Added the NFData instances to HandTypeCount and HandType
Types.