HandCounters 0.15
DataTypes: created the totHtProb function to return the total
probability of a HandType from its Count.
HandCounters: modified, from HighCard to ThreeOfAKind, count functions
in order to make sure that: no different HandTypes' instances are
counted, a present instance returns a probability of 1 and only that and
strictly better ones are returned. Rewriting countStraight because of
same worth but different instances exist (345678 and 45678, for
example); the present version is operational but half finished.