HandCounters 0.17
HandCounters: Completed the modifications in the count functions
required to make them (mostly check that they do):
always count exlusively instances of their HandType;
return a probability of 1 if an instance is already present, and, if
possible, only strictly better instances (not worth same but different);
no instances are wrongly removed in the process.
Specifically, in this commit the required changes were made to count
Flush, countFullHouse and countStraightFlush.