Analyser 1.0, DataTypes 1.0, GeneralFunctions 1.0, HandCounters 1.0, QuickCheckStuff 1.0
The first working (as originally envisioned; it was already "working")
version of the whole project!
Specific changes in this commit:

Analyser: introduced ">>" to indicate the shell; introduced commands to
analyse all the possible hands of the user and other players up to a
given round stage; improved the help string; fixed a silent bug in
RoundEnd, which was still using cardsInDeck instead of deck.
DataTypes: reordered and grouped all functions in each section;
introduced simpleTotProbs to calculate cumulative HandTypes
probabilities from a list of HandTypeCounts; removed Data.Maps and used
Prelude's lookup function instead.
GeneralFunctions: abstracted listUnZip from HandCounters (it was in a
Complete Counter).
HandCounters: made a separate section for Complete HandType instances
Counters and built a couple (and made the check functions mirror them);
made stageTcns take a Char instead of an Action as an input.
QuickCheckStuff: reflected the changes in check functions from
HandCounters and enabled single HandType checking.