GeneralFunctions 1.1, HandCounters 1.1, UnusedStuff 0.2
Played around with AbstractHandCounters, then decided not to use them
because hands would need to be generated anyway for ranking (not to say
that it would be impossible with them); will extract useful bits from
what was written so far.
GeneralFunctions: added inclusion-exclusion principle functions and a
function removing all supersets of any element.
HandCounters: made countPossHands more efficient, and also implemented
the noSupersets function above, but no difference in results was
observed...