HandCounters 1.7, DataTypes 1.1, GeneralFunctions 1.2
HandCounters: changed the method to check for FourOfAKinds in
possFullHouse (the new one is more efficient).
DataTypes: created suitsPerValue and similar functions (mirroring
suitGroups and similar) returning tuples instead of Cards, and
implemented them where appropriate in previous code in HandCounters.
General Functions: created notSubsetOf in order not to have to write
"not . a `subsetOf` b" every time, and implemented it in previous code.